Add files via upload
I added some specifics on the table limiting movement of the pieces up to the point it encounters another piece or reaches the end of the game board. I also added limits on the movement of the King since it cannot move into a position that will place it in check nor can it move into a space next to the other king (hence placing it in check again). Also, the Knight has unrestricted movement regardless if another piece is in the way and can only capture at the end of its move. You might check my wording to make sure it makes sense.